requirements.txt
pytest
pytest-cov
opencv-python
Pillow
google-cloud
